+++ composeApp/src/commonMain/kotlin/world/hachimi/app/model/GlobalStore.kt
@@ -107,6 +107,7 @@ class GlobalStore(
 
     fun updateLoudnessNormalization(enabled: Boolean) = scope.launch {
         this@GlobalStore.enableLoudnessNormalization = enabled
+        player.setReplayGainEnabled(enabled)
         dataStore.set(PreferencesKeys.SETTINGS_LOUDNESS_NORMALIZATION, enabled)
     }